Fixed search user flow
