--- todo.txt
@@ -1,57 +1,0 @@
-TO DO MVP
-
-- [x] Set up basic launch
-- [x] Find Movie DB and do basic imports
-- [x] Launch on vercel 
-- [x] Set up zustand for state
-- [x] Load TV show details 
-- [/] Load most recent shows
-- [x] Set up basic auth
-
-- [x] Set up DB 
-    - [] seed with TVDB details
-        - [] Omit news shows
-    - [x] Seed Genres
-    - [x] Seed Content Ratings 
-    - [] Finish the links to above
-    - [] Figure out cringe rating system
-    - [] Set default Cringe Rating depending on content rating
-    - [] Set up regular content update
-
-- [x] Fix carousel so it works under mobile
-    - Sort by newest release
-    - Sort by most reviewed 
-    - Sort by popular
-
-- [] - Rig up routing for: 
-    - [x] homepage 
-    - [x] Top nav with search
-    - [x] Profile page
-    - [] Individual series search 
-
-- [x] Auth set up
- - Auth0
-
-- [ ] HOMEPAGE COMPONENTS: 
-    - [] search functions
-        - [x] search bar ui
-    - [x] List of latest tv shows
-        - [] Intuit contents via rating
-    - [] - List of latest reviews
-
-- [ ] PROFILE PAGE 
-    - [x] Route by user id
-    - [x] Name, 
-    - [] Implement Bio 
-    - [] Account age
-    - [] List of reviews
-
-- [ ] FILM SELECT PAGE 
-    - [] Build UI
-    - [x] Route by ID
-    - [] Review insert section
-        - [] Show current review left by user
-    - [] Show latest reviews
-    - [] Show general consensus review weighing
-    - [] Film synopsis
-    - [] classification ratings
